backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
pydantic
email-validator
python-dotenv
python-jose[cryptography]
passlib
mercadopago
